lottocombo.localhost/backend/requirements.txt
asgiref==3.4.1
astroid==2.6.6
autopep8==1.5.7
certifi==2021.5.30
charset-normalizer==2.0.4
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.6
django-cors-headers==3.8.0
django-crontab==0.7.1
django-environ==0.4.5
django-extensions==3.1.3
djangorestframework==3.12.4
drf-yasg==1.20.0
gunicorn==20.1.0
idna==3.2
inflection==0.5.1
isort==5.9.3
itypes==1.2.0
Jinja2==3.0.1
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
mccabe==0.6.1
packaging==21.0
psycopg2-binary==2.9.1
pycodestyle==2.7.0
pylint==2.9.6
pyparsing==2.4.7
python-dotenv==0.19.0
pytz==2021.1
requests==2.26.0
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.26.6
whitenoise==5.3.0
wrapt==1.12.1
